+++ src/FolkerKinzel.VCards/md/ReleaseNotes.md
@@ -1,5 +1,6 @@
-- Fixes an issue that with `X-GENDER` and `X-WAB-GENDER` the parameters and the group name were not parsed.
-- Dependency updates
+- .NET 9.0 version of the package
+- Changes the `IDProperty(string?) ctor` to create a v7 Guid in .NET 9.0
+- Dependency update
 
 &nbsp;
 >**Project reference:** On some systems, the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and **check the "Allow" checkbox** - if it is present - in the lower right corner of the General tab in the Properties dialog.